# api/openapi.yaml
openapi: 3.0.3
info:
  title: Paladin API
  version: 1.1.0
  description: |
    S3-compatible control plane for managing object metadata and issuing pre-signed URLs
    for single PUT and multipart uploads. Includes explicit commit (complete) step.

    ## Features
    - **Metadata Management**: Track objects with labels and external references
    - **Pre-signed URLs**: Secure, time-limited upload/download URLs
    - **Multipart Uploads**: Support for large files (>5GB)
    - **Idempotency**: Safe retries with idempotency keys
    - **Multi-tenancy**: Tenant isolation via X-Tenant-ID header

    ## Versioning
    - **URL Versioning**: Major version in path (`/v1`, `/v2`)
    - **Backward Compatibility**: Minor/patch versions are backward compatible
    - **Deprecation Policy**: 6 months notice before removing deprecated features
    - **Version Header**: All responses include `X-API-Version` header

    ## CORS
    API supports CORS for browser-based clients:
    - Allowed origins configured per environment
    - Credentials supported via `Access-Control-Allow-Credentials`
    - Preflight cache: 86400 seconds (24 hours)

    ## Rate Limiting
    - Default: 1000 requests per minute per tenant
    - Headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`
    - Exceeded: HTTP 429 with `Retry-After` header
servers:
  - url: /v1
...
paths:
  /objects:
    get:
      tags: [Objects]
      summary: List objects
      operationId: listObjects
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
        - name: status
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ObjectStatus"
        - name: external_ref
          in: query
          required: false
          schema:
            type: string
        - name: created_after
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: created_before
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Objects list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListObjectsResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Objects]
      summary: Create a single object and return signed upload action (PUT)
      operationId: createObject
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateObjectRequest"
      responses:
        "201":
          description: Object created
          headers:
            Location:
              description: Canonical object resource URL
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateObjectResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /objects/{id}:
    parameters:
      - $ref: "#/components/parameters/ObjectID"

    get:
      tags: [Objects]
      summary: Get object details (without signing download URL)
      operationId: getObject
      responses:
        "200":
          description: Object details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetObjectResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    head:
      tags: [Objects]
      summary: Check object existence
      operationId: headObject
      responses:
        "200":
          description: Object exists
          headers:
            X-Object-Status:
              schema:
                $ref: "#/components/schemas/ObjectStatus"
            X-Object-Size-Bytes:
              schema:
                type: integer
                format: int64
            X-Object-Content-Type:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [Objects]
      summary: Hard delete an object (irreversible)
      description: |
        Permanently removes the object data from storage and metadata from the database.
        This operation is irreversible.
      operationId: hardDeleteObject
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204":
          description: Object deleted (or already absent)
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      tags: [Objects]
      summary: Update object status (e.g. soft delete)
      description: |
        Update the lifecycle state of an object.
        Used primarily for soft deletion (`status` = `soft_deleted`).
      operationId: updateObject
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchObjectRequest"
      responses:
        "200":
          description: Object updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ObjectCommon"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /objects/{id}/meta:
    parameters:
      - $ref: "#/components/parameters/ObjectID"

    get:
      tags: [Objects]
      summary: Get object metadata
      operationId: getObjectMeta
      responses:
        "200":
          description: Object metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetObjectMetaResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    patch:
      tags: [Objects]
      summary: Update object metadata (labels/external_ref)
      operationId: patchObjectMeta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchObjectMetaRequest"
      responses:
        "200":
          description: Updated metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetObjectMetaResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /objects/{id}/sign-upload:
    parameters:
      - $ref: "#/components/parameters/ObjectID"

    post:
      tags: [Objects]
      summary: Re-issue signed upload action for single PUT (if upload not completed)
      operationId: signObjectUpload
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignObjectUploadRequest"
      responses:
        "200":
          description: Signed upload action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignObjectUploadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /objects/{id}/sign-download:
    parameters:
      - $ref: "#/components/parameters/ObjectID"

    post:
      tags: [Objects]
      summary: Issue signed download action (GET)
      operationId: signObjectDownload
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignObjectDownloadRequest"
      responses:
        "200":
          description: Signed download action
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignObjectDownloadResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /objects/{id}/complete:
    parameters:
      - $ref: "#/components/parameters/ObjectID"

    post:
      tags: [Objects]
      summary: Commit single object upload (server validates via S3 HEAD)
      operationId: completeObject
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompleteObjectRequest"
      responses:
        "200":
          description: Object committed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CompleteObjectResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /multipart:
    post:
      tags: [Multipart]
      summary: Initiate multipart upload
      operationId: initiateMultipart
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitiateMultipartRequest"
      responses:
        "200":
          description: Multipart upload initiated
          headers:
            Location:
              description: Canonical multipart resource URL
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InitiateMultipartResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
...
  parameters:
    ObjectID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
  schemas:
    ObjectStatus:
      type: string
      enum: [pending, uploading, uploaded, complete, aborted, deleted, error]
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...