# internal/api/http/adapter.go
package httpapi

import (
	"net/http"
	"sync/atomic"

	"github.com/oleg-tkachuk/paladin/internal/config"
	"github.com/oleg-tkachuk/paladin/internal/domain"
	"github.com/oleg-tkachuk/paladin/internal/generated/api"
	"github.com/oleg-tkachuk/paladin/internal/service"

	"github.com/gin-gonic/gin"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

type OpenAPIAdapter struct {
	cfg       *config.Config
	svc       domain.ObjectsService
	hs        *service.HealthService
	started   *atomic.Bool
	version   string
	commit    string
	buildTime string
}

// NewOpenAPIAdapter creates a new OpenAPIAdapter
func NewOpenAPIAdapter(cfg *config.Config, svc domain.ObjectsService, hs *service.HealthService, started *atomic.Bool, version, commit, buildTime string) *OpenAPIAdapter {
	return &OpenAPIAdapter{
		cfg:       cfg,
		svc:       svc,
		hs:        hs,
		started:   started,
		version:   version,
		commit:    commit,
		buildTime: buildTime,
	}
}

// Ensure OpenAPIAdapter implements api.ServerInterface
var _ api.ServerInterface = (*OpenAPIAdapter)(nil)

func (s *OpenAPIAdapter) Healthz(c *gin.Context) {
	c.Status(http.StatusOK)
}

func (s *OpenAPIAdapter) HealthLivez(c *gin.Context) {
	c.JSON(http.StatusOK, api.HealthResponse{Status: "alive"})
}

func (s *OpenAPIAdapter) HealthReadyz(c *gin.Context) {
	ready, status := s.hs.CheckReady(c.Request.Context())
	healthResp := api.HealthResponse{
		Dependencies: mapDependencyStatus(status),
	}

	if !ready {
		healthResp.Status = "not_ready"
		healthResp.Reason = ptr("dependency_unavailable")
		c.JSON(http.StatusServiceUnavailable, healthResp)
		return
	}

	healthResp.Status = "ready"
	c.JSON(http.StatusOK, healthResp)
}

func (s *OpenAPIAdapter) HealthStartupz(c *gin.Context) {
	if !s.started.Load() {
		c.JSON(http.StatusServiceUnavailable, api.HealthResponse{
			Status: "starting",
			Reason: ptr("initialization_in_progress"),
		})
		return
	}
	c.JSON(http.StatusOK, api.HealthResponse{Status: "started"})
}

func mapDetailedStatus(s service.DetailedDependencyStatus) *api.DetailedDependencyStatus {
	var msg *string
	if s.Message != "" {
		msg = &s.Message
	}
	return &api.DetailedDependencyStatus{
		Status:    api.DetailedDependencyStatusStatus(s.Status),
		LatencyMs: s.LatencyMs,
		Message:   msg,
	}
}

func mapDependencyStatus(s service.DependencyStatus) *api.DependencyStatus {
	poolStats := make(map[string]map[string]interface{})
	for k, v := range s.PoolStats {
		if m, ok := v.(map[string]interface{}); ok {
			poolStats[k] = m
		}
	}

	return &api.DependencyStatus{
		Postgresql: mapDetailedStatus(s.PostgreSQL),
		Seaweedfs:  mapDetailedStatus(s.SeaweedFS),
		Breakers:   &s.Breakers,
		PoolStats:  &poolStats,
	}
}

func ptr[T any](v T) *T { return &v }

func (s *OpenAPIAdapter) Version(c *gin.Context) {
	c.JSON(http.StatusOK, api.VersionResponse{
		Service:   "paladin",
		Version:   s.version,
		GitSha:    &s.commit,
		BuildTime: &s.buildTime,
	})
}

func (s *OpenAPIAdapter) CreateObject(c *gin.Context, params api.CreateObjectParams) {
	var req api.CreateObjectRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		respondWithError(c, http.StatusBadRequest, err)
		return
	}

	ttl := 0
	if req.UploadExpiresInSeconds != nil {
		ttl = *req.UploadExpiresInSeconds
	}

	out, err := s.svc.CreateSingle(c.Request.Context(), tenantID(c), string(req.ContentType), req.SizeBytes, mapLabels(req.Labels), req.ExternalRef, ttl, params.IdempotencyKey)
	if err != nil {
		respondWithError(c, http.StatusBadRequest, err)
		return
	}

	c.JSON(http.StatusCreated, api.CreateObjectResponse{
		ObjectId:  out.ID,
		ObjectKey: out.Key,
		Bucket:    out.Bucket,
		Status:    api.Pending,
		Upload:    mapSignedAction(out.Upload),
	})
}

func (s *OpenAPIAdapter) GetObject(c *gin.Context, id openapi_types.UUID) {
	rec, err := s.svc.Get(c.Request.Context(), tenantID(c), id)
	if err != nil {
		respondWithError(c, http.StatusNotFound, err)
		return
	}

	c.JSON(http.StatusOK, mapObjectCommon(rec))
}

func (s *OpenAPIAdapter) HeadObject(c *gin.Context, id openapi_types.UUID) {
	_, err := s.svc.GetMeta(c.Request.Context(), tenantID(c), id)
	if err != nil {
		c.Status(http.StatusNotFound)
		return
	}
	c.Status(http.StatusOK)
}

func (s *OpenAPIAdapter) HardDeleteObject(c *gin.Context, id openapi_types.UUID, params api.HardDeleteObjectParams) {
	var idempotencyKey *string
	if params.IdempotencyKey != nil {
		k := string(*params.IdempotencyKey)
		idempotencyKey = &k
	}

	if err := s.svc.HardDelete(c.Request.Context(), tenantID(c), id, idempotencyKey); err != nil {
		respondWithError(c, http.StatusInternalServerError, err)
		return
	}

	c.Status(http.StatusNoContent)
}

func (s *OpenAPIAdapter) UpdateObject(c *gin.Context, id openapi_types.UUID, params api.UpdateObjectParams) {
	var req api.PatchObjectRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		respondWithError(c, http.StatusBadRequest, err)
		return
	}

	var idempotencyKey *string
	if params.IdempotencyKey != nil {
		k := string(*params.IdempotencyKey)
		idempotencyKey = &k
	}

	if err := s.svc.UpdateStatus(c.Request.Context(), tenantID(c), id, string(req.Status), idempotencyKey); err != nil {
		respondWithError(c, http.StatusInternalServerError, err)
		return
	}

	// Fetch updated object to return
	rec, err := s.svc.Get(c.Request.Context(), tenantID(c), id)
	if err != nil {
		respondWithError(c, http.StatusInternalServerError, err)
		return
	}

	c.JSON(http.StatusOK, mapObjectCommon(rec))
}

func (s *OpenAPIAdapter) CompleteObject(c *gin.Context, id openapi_types.UUID) {
	var req api.CompleteObjectRequest
	_ = c.ShouldBindJSON(&req) // Optional body

	rec, err := s.svc.CompleteObject(c.Request.Context(), tenantID(c), id, req.Etag, req.SizeBytes)
	if err != nil {
		respondWithError(c, http.StatusBadRequest, err)
		return
	}

	c.JSON(http.StatusOK, api.CompleteObjectResponse{
		Status:          mapStatus(rec.Status),
		StoredEtag:      rec.StoredETag,
		StoredSizeBytes: rec.StoredSizeBytes,
		CompletedAt:     rec.CompletedAt,
	})
}

func (s *OpenAPIAdapter) GetObjectMeta(c *gin.Context, id openapi_types.UUID) {
	rec, err := s.svc.GetMeta(c.Request.Context(), tenantID(c), id)
	if err != nil {
		respondWithError(c, http.StatusNotFound, err)
		return
	}

	c.JSON(http.StatusOK, mapObjectCommon(rec))
}

func (s *OpenAPIAdapter) PatchObjectMeta(c *gin.Context, id openapi_types.UUID) {
	var req api.PatchObjectMetaRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		respondWithError(c, http.StatusBadRequest, err)
		return
	}

	rec, err := s.svc.PatchMeta(c.Request.Context(), tenantID(c), id, mapLabels(req.Labels), req.ExternalRef)
	if err != nil {
		respondWithError(c, http.StatusBadRequest, err)
		return
	}

	c.JSON(http.StatusOK, mapObjectCommon(rec))
}

func (s *OpenAPIAdapter) ListObjects(c *gin.Context, params api.ListObjectsParams) {
	limit := 100
	if params.Limit != nil {
		limit = *params.Limit
	}
	cursor := ""
	if params.Cursor != nil {
		cursor = *params.Cursor
	}

	filter := domain.ListObjectsFilter{
		Status:      (*domain.ObjectStatus)(params.Status),
		ExternalRef: params.ExternalRef,
	}

	items, next, err := s.svc.List(c.Request.Context(), tenantID(c), filter, limit, cursor)
	if err != nil {
		respondWithError(c, http.StatusInternalServerError, err)
		return
	}

	out := make([]api.ObjectCommon, len(items))
	for i, item := range items {
		out[i] = mapObjectCommon(&item)
	}

	// Determine if there are more results
	hasMore := next != ""
	var nextCursor *string
	if hasMore {
		nextCursor = &next
	}

	resp := api.ListObjectsResponse{
		Items: out,
	}
	resp.Pagination.NextCursor = nextCursor
	resp.Pagination.HasMore = hasMore

	c.JSON(http.StatusOK, resp)
}

func (s *OpenAPIAdapter) SignObjectDownload(c *gin.Context, id openapi_types.UUID) {
	var req api.SignObjectDownloadRequest
	_ = c.ShouldBindJSON(&req)

	ttl := 0
	if req.DownloadExpiresInSeconds != nil {
		ttl = *req.DownloadExpiresInSeconds
	}

	p, err := s.svc.SignDownload(c.Request.Context(), tenantID(c), id, ttl)
	if err != nil {
		respondWithError(c, http.StatusBadRequest, err)
		return
	}

	c.JSON(http.StatusOK, api.SignObjectDownloadResponse{
		ObjectId: id,
		Download: mapSignedAction(p),
	})
}

func (s *OpenAPIAdapter) SignObjectUpload(c *gin.Context, id openapi_types.UUID) {
	var req api.SignObjectUploadRequest
	_ = c.ShouldBindJSON(&req)

	ttl := 0
	if req.UploadExpiresInSeconds != nil {
		ttl = *req.UploadExpiresInSeconds
	}

	p, err := s.svc.SignUpload(c.Request.Context(), tenantID(c), id, ttl)
	if err != nil {
		respondWithError(c, http.StatusBadRequest, err)
		return
	}

	c.JSON(http.StatusOK, api.SignObjectUploadResponse{
		ObjectId: id,
		Upload:   mapSignedAction(p),
	})
}

func (s *OpenAPIAdapter) InitiateMultipart(c *gin.Context, params api.InitiateMultipartParams) {
	var req api.InitiateMultipartRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		respondWithError(c, http.StatusBadRequest, err)
		return
	}

	ttl := 0
	if req.UploadExpiresInSeconds != nil {
		ttl = *req.UploadExpiresInSeconds
	}

	out, err := s.svc.InitiateMultipart(c.Request.Context(), tenantID(c), string(req.ContentType), req.SizeBytes, mapLabels(req.Labels), req.ExternalRef, ttl, params.IdempotencyKey)
	if err != nil {
		respondWithError(c, http.StatusBadRequest, err)
		return
	}

	c.JSON(http.StatusOK, api.InitiateMultipartResponse{
		ObjectId:  out.ObjectID,
		ObjectKey: out.ObjectKey,
		UploadId:  out.UploadID,
		PartSize:  out.PartSize,
		ExpiresAt: out.ExpiresAt,
		Bucket:    out.Bucket,
		Status:    api.Uploading,
	})
}

func (s *OpenAPIAdapter) GetMultipart(c *gin.Context, uploadId string) {
	multi, err := s.svc.GetMultipart(c.Request.Context(), tenantID(c), uploadId)
	if err != nil {
		respondWithError(c, http.StatusNotFound, err)
		return
	}

	c.JSON(http.StatusOK, api.GetMultipartResponse{
		ObjectId:  multi.ObjectID,
		ObjectKey: multi.ObjectKey,
		UploadId:  multi.UploadID,
		PartSize:  multi.PartSize,
		Bucket:    multi.Bucket,
		Status:    mapStatus(domain.ObjectStatus(multi.Status)),
		CreatedAt: multi.CreatedAt,
		UpdatedAt: multi.UpdatedAt,
	})
}

func (s *OpenAPIAdapter) AbortMultipart(c *gin.Context, uploadId string) {
	if err := s.svc.AbortMultipart(c.Request.Context(), tenantID(c), uploadId); err != nil {
		respondWithError(c, http.StatusBadRequest, err)
		return
	}

	c.JSON(http.StatusOK, api.AbortMultipartResponse{Status: api.Aborted})
}

func (s *OpenAPIAdapter) CompleteMultipart(c *gin.Context, uploadId string) {
	var req api.CompleteMultipartRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		respondWithError(c, http.StatusBadRequest, err)
		return
	}

	parts := make([]domain.CompletePart, len(req.Parts))
	for i, p := range req.Parts {
		parts[i] = domain.CompletePart{PartNumber: p.PartNumber, ETag: p.Etag}
	}

	rec, err := s.svc.CompleteMultipart(c.Request.Context(), tenantID(c), uploadId, parts)
	if err != nil {
		respondWithError(c, http.StatusBadRequest, err)
		return
	}

	c.JSON(http.StatusOK, api.CompleteMultipartResponse{
		ObjectId:        rec.ID,
		Status:          mapStatus(rec.Status),
		StoredEtag:      rec.StoredETag,
		StoredSizeBytes: rec.StoredSizeBytes,
		CompletedAt:     rec.CompletedAt,
	})
}

func (s *OpenAPIAdapter) SignPartsBatch(c *gin.Context, uploadId string) {
	var req api.SignPartsBatchRequest
	if err := c.ShouldBindJSON(&req); err != nil {
		respondWithError(c, http.StatusBadRequest, err)
		return
	}

	batch, err := s.svc.SignPartsBatch(c.Request.Context(), tenantID(c), uploadId, req.PartNumbers)
	if err != nil {
		respondWithError(c, http.StatusBadRequest, err)
		return
	}

	out := make([]api.SignPartResponse, len(batch))
	for i, b := range batch {
		out[i] = api.SignPartResponse{
			PartNumber: b.PartNumber,
			Upload:     mapSignedAction(b.Upload),
		}
	}

	c.JSON(http.StatusOK, api.SignPartsBatchResponse{
		UploadId: uploadId,
		Parts:    out,
	})
}

func (s *OpenAPIAdapter) SignPart(c *gin.Context, uploadId string, partNumber int32) {
	p, err := s.svc.SignPart(c.Request.Context(), tenantID(c), uploadId, partNumber)
	if err != nil {
		respondWithError(c, http.StatusNotFound, err)
		return
	}

	c.JSON(http.StatusOK, api.SignPartResponse{
		PartNumber: partNumber,
		Upload:     mapSignedAction(p),
	})
}

func (s *OpenAPIAdapter) GetAdminConfig(c *gin.Context) {
	// Redact sensitive fields
	resp := api.ConfigResponse{
		App: &struct {
			Env  *string `json:"env,omitempty"`
			Name *string `json:"name,omitempty"`
		}{
			Name: ptr(s.cfg.App.Name),
			Env:  ptr(s.cfg.App.Env),
		},
		Server: &struct {
			Http *struct {
				Addr               *string   `json:"addr,omitempty"`
				CorsAllowedOrigins *[]string `json:"cors_allowed_origins,omitempty"`
				ReadTimeout        *string   `json:"read_timeout,omitempty"`
				RequestIdHeader    *string   `json:"request_id_header,omitempty"`
				WriteTimeout       *string   `json:"write_timeout,omitempty"`
			} `json:"http,omitempty"`
			Mode *string `json:"mode,omitempty"`
			Name *string `json:"name,omitempty"`
		}{
			Mode: ptr(s.cfg.Server.Mode),
			Name: ptr(s.cfg.Server.Name),
			Http: &struct {
				Addr               *string   `json:"addr,omitempty"`
				CorsAllowedOrigins *[]string `json:"cors_allowed_origins,omitempty"`
				ReadTimeout        *string   `json:"read_timeout,omitempty"`
				RequestIdHeader    *string   `json:"request_id_header,omitempty"`
				WriteTimeout       *string   `json:"write_timeout,omitempty"`
			}{
				Addr:               ptr(s.cfg.Server.HTTP.Addr),
				ReadTimeout:        ptr(s.cfg.Server.HTTP.ReadTimeout.String()),
				WriteTimeout:       ptr(s.cfg.Server.HTTP.WriteTimeout.String()),
				RequestIdHeader:    ptr(s.cfg.Server.HTTP.RequestIDHeader),
				CorsAllowedOrigins: &s.cfg.Server.HTTP.CORSAllowedOrigins,
			},
		},
		Datastores: &struct {
			Postgres *struct {
				Pool *struct {
					MaxConnIdleTime *string `json:"max_conn_idle_time,omitempty"`
					MaxConnLifetime *string `json:"max_conn_lifetime,omitempty"`
					MaxConns        *int    `json:"max_conns,omitempty"`
					MinConns        *int    `json:"min_conns,omitempty"`
				} `json:"pool,omitempty"`
				Timeouts *struct {
					Connect   *string `json:"connect,omitempty"`
					Statement *string `json:"statement,omitempty"`
				} `json:"timeouts,omitempty"`
			} `json:"postgres,omitempty"`
			S3 *struct {
				Bucket         *string `json:"bucket,omitempty"`
				Endpoint       *string `json:"endpoint,omitempty"`
				ForcePathStyle *bool   `json:"force_path_style,omitempty"`
				PartSize       *string `json:"part_size,omitempty"`
				PresignTtl     *string `json:"presign_ttl,omitempty"`
				PublicEndpoint *string `json:"public_endpoint,omitempty"`
				Region         *string `json:"region,omitempty"`
				SseType        *string `json:"sse_type,omitempty"`
			} `json:"s3,omitempty"`
		}{
			Postgres: &struct {
				Pool *struct {
					MaxConnIdleTime *string `json:"max_conn_idle_time,omitempty"`
					MaxConnLifetime *string `json:"max_conn_lifetime,omitempty"`
					MaxConns        *int    `json:"max_conns,omitempty"`
					MinConns        *int    `json:"min_conns,omitempty"`
				} `json:"pool,omitempty"`
				Timeouts *struct {
					Connect   *string `json:"connect,omitempty"`
					Statement *string `json:"statement,omitempty"`
				} `json:"timeouts,omitempty"`
			}{
				Pool: &struct {
					MaxConnIdleTime *string `json:"max_conn_idle_time,omitempty"`
					MaxConnLifetime *string `json:"max_conn_lifetime,omitempty"`
					MaxConns        *int    `json:"max_conns,omitempty"`
					MinConns        *int    `json:"min_conns,omitempty"`
				}{
					MaxConns:        ptr(int(s.cfg.Datastores.Postgres.Pool.MaxConns)),
					MinConns:        ptr(int(s.cfg.Datastores.Postgres.Pool.MinConns)),
					MaxConnLifetime: ptr(s.cfg.Datastores.Postgres.Pool.MaxConnLifetime.String()),
					MaxConnIdleTime: ptr(s.cfg.Datastores.Postgres.Pool.MaxConnIdleTime.String()),
				},
				Timeouts: &struct {
					Connect   *string `json:"connect,omitempty"`
					Statement *string `json:"statement,omitempty"`
				}{
					Connect:   ptr(s.cfg.Datastores.Postgres.Timeouts.Connect.String()),
					Statement: ptr(s.cfg.Datastores.Postgres.Timeouts.Statement.String()),
				},
			},
			S3: &struct {
				Bucket         *string `json:"bucket,omitempty"`
				Endpoint       *string `json:"endpoint,omitempty"`
				ForcePathStyle *bool   `json:"force_path_style,omitempty"`
				PartSize       *string `json:"part_size,omitempty"`
				PresignTtl     *string `json:"presign_ttl,omitempty"`
				PublicEndpoint *string `json:"public_endpoint,omitempty"`
				Region         *string `json:"region,omitempty"`
				SseType        *string `json:"sse_type,omitempty"`
			}{
				Bucket:         ptr(s.cfg.Datastores.S3.Bucket),
				Region:         ptr(s.cfg.Datastores.S3.Region),
				Endpoint:       ptr(s.cfg.Datastores.S3.Endpoint),
				PublicEndpoint: ptr(s.cfg.Datastores.S3.PublicEndpoint),
				ForcePathStyle: ptr(s.cfg.Datastores.S3.ForcePathStyle),
				PresignTtl:     ptr(s.cfg.Datastores.S3.PresignTTL.String()),
				PartSize:       ptr(s.cfg.Datastores.S3.PartSizeRaw),
				SseType:        ptr(s.cfg.Datastores.S3.SSEType),
			},
		},
		Policy: &struct {
			AllowedContentTypes *[]string `json:"allowed_content_types,omitempty"`
			MaxMultipartSize    *string   `json:"max_multipart_size,omitempty"`
			MaxObjectSize       *string   `json:"max_object_size,omitempty"`
			MaxPartSize         *string   `json:"max_part_size,omitempty"`
			MinPartSize         *string   `json:"min_part_size,omitempty"`
			PresignGetTtl       *string   `json:"presign_get_ttl,omitempty"`
			PresignPutTtl       *string   `json:"presign_put_ttl,omitempty"`
		}{
			MaxObjectSize:       ptr(s.cfg.Policy.MaxObjectSizeRaw),
			MaxMultipartSize:    ptr(s.cfg.Policy.MaxMultipartSizeRaw),
			MinPartSize:         ptr(s.cfg.Policy.MinPartSizeRaw),
			MaxPartSize:         ptr(s.cfg.Policy.MaxPartSizeRaw),
			PresignPutTtl:       ptr(s.cfg.Policy.PresignPutTTL.String()),
			PresignGetTtl:       ptr(s.cfg.Policy.PresignGetTTL.String()),
			AllowedContentTypes: &s.cfg.Policy.AllowedContentTypes,
		},
		Security: &struct {
			EnableRls                *bool `json:"enable_rls,omitempty"`
			RejectTenantMismatch     *bool `json:"reject_tenant_mismatch,omitempty"`
			TrustTenantIdFromRequest *bool `json:"trust_tenant_id_from_request,omitempty"`
		}{
			TrustTenantIdFromRequest: ptr(s.cfg.Security.TrustTenantIDFromRequest),
			RejectTenantMismatch:     ptr(s.cfg.Security.RejectTenantMismatch),
			EnableRls:                ptr(s.cfg.Security.EnableRLS),
		},
		Housekeeping: &struct {
			EnableReaper *bool   `json:"enable_reaper,omitempty"`
			GcInterval   *string `json:"gc_interval,omitempty"`
			MultipartTtl *string `json:"multipart_ttl,omitempty"`
			PendingTtl   *string `json:"pending_ttl,omitempty"`
		}{
			EnableReaper: ptr(s.cfg.Housekeeping.EnableReaper),
			PendingTtl:   ptr(s.cfg.Housekeeping.PendingTTL.String()),
			MultipartTtl: ptr(s.cfg.Housekeeping.MultipartTTL.String()),
			GcInterval:   ptr(s.cfg.Housekeeping.GCInterval.String()),
		},
		RateLimit: &struct {
			Burst             *int     `json:"burst,omitempty"`
			MaxTenants        *int     `json:"max_tenants,omitempty"`
			RequestsPerSecond *float32 `json:"requests_per_second,omitempty"`
		}{
			RequestsPerSecond: ptr(float32(s.cfg.RateLimit.RequestsPerSecond)),
			Burst:             ptr(s.cfg.RateLimit.Burst),
			MaxTenants:        ptr(s.cfg.RateLimit.MaxTenants),
		},
		Cache: &struct {
			Enabled *bool   `json:"enabled,omitempty"`
			MaxSize *int    `json:"max_size,omitempty"`
			Ttl     *string `json:"ttl,omitempty"`
		}{
			Enabled: ptr(s.cfg.Cache.Enabled),
			MaxSize: ptr(s.cfg.Cache.MaxSize),
			Ttl:     ptr(s.cfg.Cache.TTL.String()),
		},
		Timeouts: &struct {
			DefaultOperation *string `json:"default_operation,omitempty"`
			FastOperation    *string `json:"fast_operation,omitempty"`
			LongOperation    *string `json:"long_operation,omitempty"`
			S3Operation      *string `json:"s3_operation,omitempty"`
		}{
			FastOperation:    ptr(s.cfg.Timeouts.FastOperation.String()),
			DefaultOperation: ptr(s.cfg.Timeouts.DefaultOperation.String()),
			S3Operation:      ptr(s.cfg.Timeouts.S3Operation.String()),
			LongOperation:    ptr(s.cfg.Timeouts.LongOperation.String()),
		},
		Idempotency: &struct {
			Enabled *bool   `json:"enabled,omitempty"`
			Ttl     *string `json:"ttl,omitempty"`
		}{
			Enabled: ptr(s.cfg.Idempotency.Enabled),
			Ttl:     ptr(s.cfg.Idempotency.TTL.String()),
		},
		Otel: &struct {
			Enabled  *bool   `json:"enabled,omitempty"`
			Endpoint *string `json:"endpoint,omitempty"`
			Insecure *bool   `json:"insecure,omitempty"`
			Protocol *string `json:"protocol,omitempty"`
		}{
			Enabled:  ptr(s.cfg.OTel.Enabled),
			Endpoint: ptr(s.cfg.OTel.Endpoint),
			Protocol: ptr(s.cfg.OTel.Protocol),
			Insecure: ptr(s.cfg.OTel.Insecure),
		},
	}

	c.JSON(http.StatusOK, resp)
}

func mapSignedAction(p domain.Presigned) api.SignedAction {
	return api.SignedAction{
		Url:       p.URL,
		Method:    api.SignedActionMethod(p.Method),
		Headers:   &p.Headers,
		ExpiresAt: p.ExpiresAt,
	}
}

func mapStatus(s domain.ObjectStatus) api.ObjectStatus {
	return api.ObjectStatus(s)
}

func mapLabels(l *api.Labels) map[string]string {
	if l == nil {
		return nil
	}
	return (map[string]string)(*l)
}

func mapObjectCommon(rec *domain.Object) api.ObjectCommon {
	labels := api.Labels(rec.Labels)
	return api.ObjectCommon{
		ObjectId:        rec.ID,
		ObjectKey:       rec.ObjectKey,
		Bucket:          rec.Bucket,
		ContentType:     rec.ContentType,
		SizeBytes:       rec.SizeBytes,
		Status:          mapStatus(rec.Status),
		Labels:          &labels,
		ExternalRef:     rec.ExternalRef,
		CreatedAt:       rec.CreatedAt,
		UpdatedAt:       rec.UpdatedAt,
		CompletedAt:     rec.CompletedAt,
		DeletedAt:       rec.DeletedAt,
		StoredEtag:      rec.StoredETag,
		StoredSizeBytes: rec.StoredSizeBytes,
	}
}